refactor: Apply fixes based on AI review feedback (#943)
- Extract pure utilities to _utils/ for testability (URL building, priority calculation)
- Add Zod schema validation for workbook placements
- Refactor KanbanBoard/Column/TabBar components with Svelte 5 Runes
- Update auth rule to document validateAdminAccess() usage constraints
- Add AI review lessons and refactoring notes to dev-notes

+++ src/features/workbooks/zod/schema.ts
@@ -64,8 +64,8 @@ export const workBookSchema = z.object({
 
 export const workBookPlacementSchema = z
   .object({
-    id: z.number().positive(),
-    priority: z.number().positive(),
+    id: z.number().int().positive(),
+    priority: z.number().int().positive(),
     taskGrade: z.nativeEnum(TaskGrade).nullable(),
     solutionCategory: z.nativeEnum(SolutionCategory).nullable(),
   })
